Refactor synth to module directory.
